feat(webapp): responsables via l'objet détail (panier + indicateur)
