+++ apps/mb-webapp/src/api/paniers.ts
@@ -10,10 +10,6 @@ import {
   panierContactsUtilesApiModelSchema,
 } from '@pilote/mb-shared/panierContactUtile'
 import {
-  type PanierResponsablesApiModel,
-  panierResponsablesApiModelSchema,
-} from '@pilote/mb-shared/panierResponsable'
-import {
   type PanierTauxProgressionApiModel,
   panierTauxProgressionApiModelSchema,
 } from '@pilote/mb-shared/panierTauxProgression'
@@ -41,13 +37,6 @@ export const fetchPanierTauxProgression = async ({
     .get(`paniers/${panierId}/taux-progression`, { searchParams: { individu } })
     .json()
   return panierTauxProgressionApiModelSchema.parse(json)
-}
-
-export const fetchPanierResponsables = async (
-  panierId: string,
-): Promise<PanierResponsablesApiModel> => {
-  const json = await apiClient.get(`paniers/${panierId}/responsables`).json()
-  return panierResponsablesApiModelSchema.parse(json)
 }
 
 export const fetchPanierContactsUtiles = async (
